Preserve decks built while user is not logged in
Basically this:

- Serializes your deck-in-progress every time you make an edit and
tosses that value into session storage
- Whenever you load the deck builder page, we first check session
storage to see if anything is there. If something is there, it gets
loaded into the deck builder
- Logged out users now see a "log in to save your deck" type button
instead of the usual save when on the deck builder page.
- The BE will now attempt to return you to the same page you were on
after you log in.

This will require updates after we layer in deck editing capabilities.
E.g. we may want to ignore edits in session storage if the deck ID for
those edits doesn't match the deck you're trying to edit now.

I added sessions support to the express server that wraps our Next app.
This allows us to flash the redirect URL and get it back after bouncing
off the auth provider. If feels a little heavy handed to add sessions
just for this... I'm open to other suggestions. Other common solutions
include flashing to a cookie or DB storage - neither of those felt
particularly appealing either.