Add static events page content (#103)
I also renamed the page from "Tournaments" to "Events" to match the
designs.

Why did I not do a complete refactor to switch tournaments --> events? I
did :D and then realized having a bunch of JavaScript variables called
`Event` and `event` either clobbered things or was really confusing. So
now we outwardly refer to things as "Events" but all the internal
references are to "Tournaments"... hopefully that doesn't make anyone
too sad :(.